requirements.txt
python=3.9
pymongo
python-dateutil
watchdog
